Merged master's large compliance update (DDR-007, compliance audits, hooks,
architecture tests) with refactoring branch's HIGHLANDER resource value work.

Key conflict resolutions:
- arc42/08: Combined HIGHLANDER (8.22) and DDR-007 (8.23) sections
- Emergency system: Using ActiveEmergencyState pattern with sync methods
- Token system: Restored Dictionary<ConnectionType, int> return type
- Resource facades: Kept async patterns with ApplyConsequence
- Added CompoundRequirement.CreateForConsequence factory method

Both branches' changes preserved:
- Master: DDR-007 compliance, audit reports, architecture tests, hooks
- Refactoring: HIGHLANDER resource value unification, Consequence class

+++ src/GameState/RouteRecommendation.cs
@@ -7,7 +7,7 @@ public class RouteRecommendation
 {
     public RouteOption RecommendedRoute { get; set; }
     public string Justification { get; set; }
-    public double EfficiencyScore { get; set; }
+    public int EfficiencyScore { get; set; }
     public OptimizationStrategy Strategy { get; set; }
     public string ResourceAnalysis { get; set; }
     public string AlternativeOptions { get; set; }
